+++ oinkbrew/StatusMessage.cpp
@@ -48,9 +48,9 @@ void StatusMessage::send() {
     jsonMessage.concat(Spark.deviceID().c_str());
     jsonMessage.concat("\",\"datetime\":");
     jsonMessage.concat(Time.now());
-    jsonMessage.concat(",\"device_mode\":\"");
-    jsonMessage.concat(reinterpret_cast<const char*>(sparkInfo.mode));
-    jsonMessage.concat("\", \"firmware_version\":\"");
+    jsonMessage.concat(",\"device_mode\":");
+    jsonMessage.concat(sparkInfo.mode);
+    jsonMessage.concat(", \"firmware_version\":\"");
     jsonMessage.concat(OINK_BREW_VERSION);
     jsonMessage.concat("\",\"ip_address\":\"");
     jsonMessage.concat(Helper::getLocalIpStr().c_str());