+++ oinkbrew/TcpLogger.cpp
@@ -54,7 +54,7 @@ void TcpLogger::logDeviceValues() {
 	IPAddress oinkWebIp(sparkInfo.oinkWeb);
 
 	// only log if device is not in MANUAL mode
-	if (strcmp(reinterpret_cast<const char*>(sparkInfo.mode), "MANUAL") == 0 || oinkWebIp == IPAddress(0, 0, 0, 0))
+	if (sparkInfo.mode == SPARK_MODE_MANUAL || oinkWebIp == IPAddress(0, 0, 0, 0))
 		return;
 
 	request.ip = oinkWebIp;