+++ oinkbrew/devices/DeviceManager.h
@@ -40,7 +40,8 @@ struct DeviceRequest {
 	uint8_t pin_nr;
 	DeviceAddress hw_address;
 	bool is_invert;
-	uint8_t value;
+	float value;
+	float offset;
 };
 
 
@@ -65,6 +66,7 @@ public:
 	static void sendDevice(TCPClient& client, DeviceRequest& deviceRequest);
 	static void searchAndSendDeviceList(TCPClient& client);
 	static void toggleActuator(DeviceRequest& deviceRequest, char* response);
+	static void setOffset(DeviceRequest& deviceRequest);
 private:
 	static void processActuators(Device devices[], ActiveDevice activeDevices[], uint8_t& slot);
 	static void processOneWire(Device devices[], ActiveDevice activeDevices[], uint8_t& slot);